Return user entry even if it's not a valid one
This commit updates the default condition of the getHumanChoice()
function. After alerting the user that they have entered an invalid game
keyword, the function will still return the entered string.

This change allows the function to send a string to the playRound()
function to print out in case the user's entry doesn't match any valid
game choice.