Complete playGame() logic to play 5 rounds
This commit completes the playGame() function logic for playing 5 rounds
of rock-paper-scissors.

I have removed required parameters/arguments for the playRound() function.
I moved the humanSelection and computerSelection variables to be
under the playRound() function. Calling playRound() will now prompt for
a new value for humanSelection as well as generate a new computerSelection
in order to play the new round.

The game logic is sound to the best of my understanding.